Support SslContextBuilder customizer when configuring TLS with ServerBuilder
Motivation:
When configuring a TLS server, the server needs to configure some common part, for example, setting a `SslProvider` to the context depending on the `Flags.useOpenSsl()` property. But it should be done by a user implementation if the user builds a `SslContext` by themselves. It would be better provide a way to customize the `SslContext` after configuring common part by Armeria.

Modifications:
- Added the following methods to the `ServerBuilder` and `AbstractVirtualHostBuilder`:
  - `tls(File keyCertChainFile, File keyFile, Consumer<SslContextBuilder> tlsCustomizer)`
  - `tls(File keyCertChainFile, File keyFile, @Nullable String keyPassword, Consumer<SslContextBuilder> tlsCustomizer)`
  - `tls(KeyManagerFactory keyManagerFactory, Consumer<SslContextBuilder> tlsCustomizer)`
- Used `tls(KeyManagerFactory, Consumer<SslContextBuilder>)` when configuring TLS server with `spring-boot-autoconfigure` and `spring-boot-webflux-autoconfigure` module.

Result:
- Hopefully closes #1711.