Add options to customize the JSON marshalling behavior of gRPC service/client (#1753)
Fixes #1696

Motivation:

Currently the GrpcService initialisation via the builder doesn't allow the option of preserving Proto field names in the output. If we enable the default constructor of the GrpcService to allow the option of setting preservingProtoFieldNames then the same can be passed onto GrpcJsonUtil. This would be helpful if we don't need to modify every Proto file to include json_name option.

Modifications:

- Add options to customize the JSON marshalling behavior of gRPC service/client