Log response bytes transferred when sending a server timeout response… (#1775)
* Log response bytes transferred when sending a server timeout response, and don't try logging response bytes when a client timed out, therefore no response was sent.

Currently, if the server times out a request, it does send a timeout response, but does not call the log handler to indicate it sent bytes. Also, both tracing implementations try to log response bytes timing always, but if the client times out a request, there was never a response so this needs to actually be checked.