Allow non-armeria trace context for clients which may be outside an armeria server. (#1768)
While it's never a good idea to use the default CurrentTraceContext for an armeria server, I think it can be for an armeria client which is being used outside of an armeria server in a blocking way. Though I could be missing a corner case too since tracing is hard.