Consolidate SSL logic and enable TLSv1.3 by default. (#1779)
Motivation:

While working on other SSL stuff, I noticed that the logic for SSL configuration is in two places and a bit confusing. And it's not too hard to enable TLSv1.3, so figured may as well turn it on since it seems to be getting more widely used now. FWIU, openssl intends users to automatically get it enabled when updating their dependency version, so figure similar for Armeria should be acceptable. For reference, https://wiki.openssl.org/index.php/TLS1.3

Modifications:

- Extracted SSL logic into `SslContextUtil`
- Allowed a user to customize TLS cipher suite (with validation)