Preserve percent-encoded character after parsing URI by PathAndQuery
Motivation:
Currently, percent-encoded characters only for ampersand(`&`), semi-colon(`;`) and equal(`=`) are preserved
after parsing URI by `PathAndQuery`. So, if a request has a percent-encoded `#`(`%23`), it is decoded
as a raw `#` character. Due to this behavior, `QueryStringDecoder` cannot parse parameters correctly
if a query string has a `#`, because `#` is a fragment identifier and the decoder ignores everything
after the identifier.

Modifications:
- Every percent-encoded characters for reserved characters are preserved after parsing URI by `PathAndQuery`.
- Reserved characters. (from `https://tools.ietf.org/html/rfc3986#section-2.2`)
  - `:` / `/` / `?` / `#` / `[` / `]` / `@`
  - `!` / `$` / `&` / `'` / `(` / `)` / `*` / `+` / `,` / `;` / `=`