Rename HttpData.of to HttpData.wrap, add copyOf, and deprecate offset. (#1797)
Motivation:

Currently, we allow creating a `HttpData` as a slice of a `byte[]`. However, in practice we almost never use it, only when constructing an `HttpData` after decoding a `CharSequence` (not `String`) which is very rare usage, so I've changed it to copy out the slice. We previously also did so when retrieving the result of Thrift serialization, but replaced that with using `ByteBufHttpData`. So instead, we can change the slicing APIs to use `ByteBuf` and remove the need for users to worry about `offset` / `length` when accessing the data array.

Without the ability to slice `HttpData`, a user can access the array of an `HttpData` for processing by simply calling `.array()`, where it has been easy to forget to call `.offset()` and `.length()` as is currently required. It's especially problematic because it still usually works, but couldn't be guaranteed to work in all situations.

Modifications:

* Deprecate `offset()` and add `HttpData.wrap()` and `HttpData.copyOf()`
* Remove usage of offset
* Clarify Javadocs for `byte[]` factories.
* Add wrapping implementation

Result:

- Fixes #1771