Add a utility for executing junit5 tests with timeout and convert a parameterized test to it (#1802)
Was curious if JUnit 5 has the same problem as #1601 and it doesn't! So I want to migrate all the parameterized tests to JUnit 5 (or at least the gRPC ones since I use them a lot and it's annoying that I can't use Gradle runner because of them).

I found that the biggest missing piece is global timeouts. 5.5 will have it, but that will probably take a while to be released. Also, I don't think they'll support disable-on-debug in their first release either. So wondering if it makes sense to add this utility for running a lambda with a timeout using the current JUnit 5 style.