Uses log object consistently during server handling
We are currently working on a way to do secondary sampling for scenarios
such as customer support. For example, 100% data for tracing with a
specific device or customer ID (ex investigate Adrian's stickers). This
is cleaner to implement when there's only one type parsed when making a
trace context.

See https://github.com/openzipkin/brave/pull/986