Implement `ProtocolDetectionHandler.exceptionCaught()`
Motivation:

When an Armeria server is serving multiple protocols on the same port,
the last `ChannelHandler` in a `ChannelPipeline` is a `ProtocolDetectionHandler`,
which does not implement `exceptionCaught()` handler method properly.

As a result, a user will see the following warning message when the
connection is closed before the current session protocol is detected.

    An exceptionCaught() event was fired, and it reached at the tail of
    the pipeline. It usually means the last handler in the pipeline did
    not handle the exception.

Modifications:

- Implement `ProtocolDetectionHandler.exceptionCaught()` properly.
- Make sure the connection is closed when protocol detection is not
  finished in time.

Result:

- No more confusing warning log message on premature disconnection.
- Timeout is now enforced during protocol detection.
- Fixes #2043