Always use `ServiceRequestContext` on the server side
Motivation:

`HttpServerHandler` creates an `EarlyRespondingRequestContext` which
does not implement `ServiceRequestContext` when handling an invalid
request. The default `AccessLogWriter` implementations expects all
`RequestContext`s on the server side are `ServiceRequestContext`. As a
result, an invalid request triggers a `ClassCastException` in
`AccessLogWriter`.

Such an assumption that a service-side context is always
`ServiceRequestContext` is reasonable, so we should always provide a
`ServiceRequestContext` rather than just a `RequestContext`.

Modifications:

- Remove `EarlyRespondingRequestContext`.
- Always use `DefaultServiceRequestContext`, even for invalid requests
  and the `OPTIONS *` request.
- Enable access logs for the test cases that tests invalid paths.
- Simplify the response generation logic in `HttpServerHandler`.
- Simplify how `HttpServerHandler` retrieves the current `SSLSession`.

Result:

- `AccessLogWriter` always sees a `ServiceRequestContext`.