Fix noisy `WARN` logs on client disconnection
Related: #2248
Motivation:

When a client disconnects from a server, the `HttpResponse` being
prepared by a service was simply aborted with an `AbortedStreamException`.

However, since 0.96.0, thanks to #2248, an `HttpResponse` is aborted
with more specific exception, such as:

- `ClosedSessionException` on a disconnection
- `Http2Exception.StreamException` on an HTTP/2 RST_STREAM frame

However, `HttpResponseSubscriber` was not updated to handle the above
exceptions, which means they will leave noisy `WARN` level logs.

Modifications:

- Handle `ClosedSessionException` and `StreamException` properly.
- Make sure to cancel the health check response timeout task when a
  client disconnected, so that the timeout task do not stay in the event
  loop task queue longer than necessary.

Result:

- Less noisy logging
- Tad bit more efficient long-polling health check