Return a CompletableFuture from API methods that checks whether… (#2275)
… blocking methods are called from an event loop or a non-blocking thread.

We have heard many stories of bad experiences because of blocking the event loop in an armeria server request. While we wouldn't be able to check all cases of this (e.g., calling JDBC from an event loop), it's fairly simple to check this for futures returned by Armeria methods and I think this covers a lot of cases.  This changes to return a special implementation of `CompletableFuture` that logs from blocking methods with a stack trace.