Fix a bug where causes cache misses from RouteCache
Motivation:
While I was making #2102, I fixed `hashCode()` and `equals()` implementation
of `DefaultRoutingContext` class as using `query()` and `headers()` rather
than `contentType()` and `acceptTypes()`. It seemed okay because I added
the properties `params()` and `headers()` to `RoutingContext`, but it was
a mistake from the `RouteCache`'s view. It would cause cache misses from
`RouteCache` because the variation of cache key becomes wide.

Modifications:
- Use `contentType()` and `acceptTypes()` from `hashCode()` and `equals()`
  of `DefaultRoutingContext` class.

Result:
- `RouteCache` can cache routes as the same as before.