Add content previewing client and service decorators Motivation: Related #2189
We currently, store `HttpData`s to the `ContentPreviewer` right before writing to the socket to preview the content.
However, this makes the server decompress twice when there's a `EncodingService` and a user wants to preview decompressed content.
It's also, not the way we do for this kinds of thing which is intercepting the requests and responses and manipulating.
We use decorators for that.

Modifications:
- Add `ContentPreviewingClient` and `ContentPreviewingService`

Result:
- Close #2189