Fix a bug where callbacks are invoked with the wrong context (#2415)
Motivation:
When we send a data, we don't buffer but flush them to the socket one by one.
However, under heavy load, especially when using HTTP/2, there's a chance that the data is not fully written(partial write) or nor written at all because of the lack of the space in sending buffer. http://man7.org/linux/man-pages/man2/write.2.html
Even though we try to push the data to the socket while spinning (16 times by default), it could fail.
In that case, we just store data to `ChannelOutboundBuffer` and try it next time when writing other data.

This is where the trouble begins. The next write that would be handling a different request, which means different `RequestContext` in the thread-local, completes the writing and executes all callbacks. Some of the callback should not be invoked with the `RequestContext`.

This could happen to some of the other `ChannelFuture`s used, so I updated them as well.

Modification:
- Remove the `RequestContext` if exists in `ChannelFuture` listener.

Result:
- Callbacks are not invoked with the wrong context in the thread-local
- Close #2414