Rename debug.proto to more internal-sounding type (#2441)
I found that unfortunately the gRPC outer class still has some, uncommonly used, functionality (holds the `Descriptor` of the proto) so couldn't remove it completely. Renamed it to include the word `Internal` (the public apis in the class are useless, just the descriptor specifications stored there are what we still want to keep) and excluded it from javadoc, hope this is ok.