Make FailFastException public
Motication:

User got a trouble on make test code when (s)he tried to do a test on
circuitBreaker and when it raises an exception (FailFastException).
Originally, to reproduce the exception, user needs to use Reflection to
make exception.

Modification:

Make FailFastException consturctor public, add java doc comment for it.
Make parameter checking for nonNull

Result:

public constructor for FailFastException and java doc