Allow `RequestContextExporter` to export an `AttributeKey` with different aliases
Motivation:
When a user tries to export an attribute with different aliases, the later one will be ignored.
For example:
```xml
<export>attrs.request_id:com.example.Context#KEY:com.example.ContextRequestIdString</export>
<export>attrs.trace_id:com.example.Context#KEY:com.example.ContextTraceIdString</export>
```
The `attrs.trace_id` is ignored and only the `attrs.request_id` is output.
Because the `ExportEntry` overrides `equals` that only compares `ExportEntry.key`.  When it is added to `RequestContextExporterBuilder
.attrs`, the later `ExportEntry` is ignored.

Modifications:
- Change `ExportEntry.equals` to compare `key` and `exportKey`

Result:
You can now export an `Attribute` with different aliases.
Fixes: The second issue in https://github.com/line/armeria/issues/2513#issuecomment-588593453