Fix a bug where response is not closed with an exception (#2590)
Motivation:
When an HTTP/2 stream is closed by the other side before sending a request headers, we should call `response.close(cause)` to close the response.
If we don't call that, the request is closed later by the peer's timeout mechanism because the response timeout and idle timeout is not activated.

Modification:
- Call `response.close(cause)` when HTTP/2 stream is closed and fails to send a request header.

Result:
- You no longer see the dangling unfinished responses.