Fix a bug where a connection is not close immediatley when HTTP/2 PING write fail
Motivation:

If a connection is closed unexpectedly by network shutdown,
the opposite does not receive close signals such as GOAWAY frame and
the connection is alive until cleaning up by request timeout, idle timeout, or PING fail.

When PING writes fail, KeepAliveHandler tries to close the connection, which takes for the idle timeout.
Because it is configured as `gracefulShutdownTimeoutMillis` for Netty Http2ConnectionHandler.
https://github.com/line/armeria/blob/fec9fb833275eba59eb9afaf25fc8b7c28132f80/core/src/main/java/com/linecorp/armeria/server/Http2ServerConnectionHandler.java#L56-L62

Modifications:

* Close a connection immediately if KeepAliveHandler is closing.

Result:
Clean up a dead connection immediately when PING write fails.