Allow injecting `null` into annotated service methods if annotated with `@Nullable`
Motivation:

Currently, sending a request without all query parameters will fail with
a `400 Bad Request` error response. For example, sending a `GET /get`
request will fail with a `400 Bad Request` response because the query
parameter `foo` does not exist:

    public class MyService {
        @Get("/get")
        public String get(@Param String foo) {
            return foo;
        }
    }

To make the parameter `foo` optional, a user has to annotate it with
`@Default` annotation or change its type to `Optional<String>`.

It would be nice if we also support `@Nullable` annotation, so that a
user can tell that `null` (missing value) is allowed for injection.

    public class MyService {
        @Get("/get")
        public String get(@Param @Nullable String foo) {
            //                   ^^^^^^^^^
            return foo;
        }
    }

Modifications:

- Allow injecting `null` if the parameter is annotated with `@Nullable`.
- Update documentation

Result:

- A simple way to handle an optional parameter/header.
- Closes #2766