Revert the `Pooled*` API
Related: #2859 #2862 #2863 #2448 #1938 #1936

Motivation:

The new pooled API introduced in #2448 helped us to hide the API related
with pooled Netty `ByteBuf`s so that an advanced user has an access to
them only when the user really meant to.

However, it introduced a few drawbacks at the same time:

- There's an overhead where a non-pooled `HttpData` must be converted
  into a `PooledHttpData` when consuming via the pooled API. The current
  implementation also wraps an existing `StreamMessage` with a
  `FilteredStreamMessage`, making an extra level of indirection for
  every element emitted.
- It is not simple enough to implement `subscribeWithPooledObjects()` or
  `aggregateWithPooledObjects()` operation optimized for a specific
  `StreamMessage` implementation.

The overhead may be reduced to some extent by some optimizations
mentioned in the previous discussions, but it would take non-trivial
effort with increased complexity in type hierarchy and implementation.
I'm not sure we're taking the right balance between user experience and
developer experience and it sometimes may be confusing even to users.

Also, given that the advanced users usually look into this feature for
higher throughput and less footprint, I'm not sure if it's a good idea
to introduce extra overhead just for API isolation.

In hindsight, I should have been more cautious about the potential
overhead and complexity of the new API before getting excited about it.
I'm sorry about what's happening in this changeset.

Modifications:

- Revert the new pooled API introduced in #2448 except the following:
  - `PooledHttpData` will not be removed.
  - The `InputStream` returned by `ByteBufHttpData.toInputStream()` will
    still not increase the reference count of the underlying buffer.
- Additional changes:
  - `ByteBufHttpData`
    - Not public anymore
    - Overall clean-up and improvements with a test case
  - Fixed a bug where `HttpMessageAggregator` generates `HttpData`
    sometimes with and sometimes without EoS flag.
  - Workaround for `address already in use` bind errors during ZooKeeper
    tests

Result:

- Performance is now better, even on a simple unary gRPC benchmark. I
  expect the improvements will be greater when handling streams without
  manipulation, e.g. proxy servers.
  - Before revert (b5289544cc6ef07e35c1349d320d420e27fa726c):

        Benchmark                         (clientType)   Mode  Cnt    Score   Error  Units
        DownstreamSimpleBenchmark.simple        NORMAL  thrpt  100  950.792 ± 3.792  ops/s

  - After revert (this commit):

        Benchmark                         (clientType)   Mode  Cnt    Score   Error  Units
        DownstreamSimpleBenchmark.simple        NORMAL  thrpt  100  968.421 ± 3.792  ops/s