Add `CheckReturnValue` annotation on streaming output
Motivation:

While fixing various buffer leaks in our code, I found most leaks were
occurring because we forgot to subscribe on the `StreamMessage` objects
returned. We could add the `@CheckReturnValue` annotation to the methods
that return a `StreamMessage` so that errorprone users gets an error in
such cases.

Modifications:

- Annotate stream-returning methods with `@CheckReturnValue`.

Result:

- Less mistakes in the future