Fix not to raise an exception from `(Service|Client)RequestContext.currentOrNull()` Motivation: `ServiceRequestContext.currentOrNull()` raises an `IllegalArgumentException` if the current ctx is `ClientRequestContext` and `ctx.root()` returns `null`.
It's very confusing because all other APIs returns `null` if the name ends with `orNull`.
So I think the API should return `null` instead of raising an exception.
We could use `ClientRequestContext.roo()` to find out if the client call is invoked by a server request or not.

Modification:
- Returns `null` instread of raising an exception from `(Service|Client)RequestContext.currentOrNull()`

Result:
- `(Service|Client)RequestContext.currentOrNull()` does not throw an exception anymore.