Fix a bug where gRPC and Thrift service name is not overriden by defaultServcieName
Motivation:

The service name and method of gRPC and Thrift are set automatically to `RequestLog.*name()` by gRPC and Thrift service.
However, the `defaultServiceName` which was specified when building a service should have a higher priority.
By setting gRPC and Thrift service when a request side is ended,
all service types will have a consistent behavior with regards to `RequestLog.serviceName()` and `ReqeustLog.name()`.

Modifications:

- Set `RequestLog.serviceName()` that was derived from a service stub when `defaultServiceName` is not configured.
- Set `RequestLog.name()` that was derived from a method of service stub when `defaultLogName` is not configured.

Result:

gRPC and Thrift service now respect a user defined `defaultServiceName` and `defaultLogName`.