Fix a bug where metrics are not collected when `metricsPath` is set to empty in Spring Boot
Motivation:

Currently, `MetricCollectingService` is not registered as a decorator when `metricsPath` is set to empty.
https://github.com/line/armeria/blob/e872d75a03de3bc384b7416d16c82dfd73187ccb/spring/boot2-autoconfigure/src/main/java/com/linecorp/armeria/internal/spring/ArmeriaConfigurationUtil.java#L109
From the original specification, this is not a expected behavior.
https://github.com/line/armeria/blob/fbe90f0f2a952578012a552c26b3f4f8bdb33c8c/spring/boot2-autoconfigure/src/main/java/com/linecorp/armeria/spring/ArmeriaSettings.java#L279-L285

Users might want to hide the metric path from the public port and expose it only via the management port of Spring Actuator.

Modification:

- Bind a `MetricCollectingService` decorator when `eanbleMetrics` is true.

Result:

- Armeria Spring Server collects metrics correctly even `ArmeriaSettins.metricPath` is set to empty string.