Allow to set a custom route for gRPC service
Motivation:

When users want to serve unframed gRPC service, sometimes they might want to use RESTful URL for the endpoint.
A gRPC service path is not a directory which is not divided by `/`.
It is the same as package name such as "/api.v1.service.UserService/GetUsers".
It is not natual for REST client. "/api/v1/service/users/GetUsers" would be better for security control or logging.

See #3072 for more details.

Modifications:

- Add `addService(path, GrpcService)` to `GrpcServiceBuilder`
- Make HandlerRegistry.services() return pairs of route path and `ServerServiceDefinition`.

Result:

- You can now bind a gRPC service to a custom route path.
- Fixes #3072