Fix a bug where a gRPC ServerCall is invoked after a stream is closed
Motivation:

When a stream is closed by a remote peer, `ArmeriaServerCall` is closed by `res.whenComplete()` listener.
https://github.com/line/armeria/blob/6e24e7817f8be94ab21ddba7879e9f84ca840229/grpc/src/main/java/com/linecorp/armeria/server/grpc/ArmeriaServerCall.java#L173-L182

If a service performs a long running task, `ArmeriaServerCall.sendHeaders()` or
`ArmeriaServerCall.sendMessage()` will be called when a response is written.
Because gRPC-Java `ServerCallStreamObserver` implementation invokes `ServerCall` even if a call is cancelled with some conditions.
https://github.com/grpc/grpc-java/blob/d52b359631ee1ac46f47d72d1015e1dd6ce9312c/stub/src/main/java/io/grpc/stub/ServerCalls.java#L346-L358

Modifications:

- Check `cancelled` flag before validating the closing states.

Result:

- `ArmeriaServerCall` is now closed exatly once.