Add Accept-Language parsing and prioritizing
Motivation:

Accept-Language is a client supplied header detailing what locales the
client user prefers and the weight of the preference. It's really
mostly useful to detect the default language of the site.

Modifications:

Added method `acceptLanguages()` and `getLocale(supportedLocales)` to
`RoutingContext` and all it's subclasses.

Add a lazy init property for `acceptLanguages()` to avoid parsing
unless needed.

Result:

Closes #3179