Add Accept-Language parsing and prioritizing (#3177)
Motivation:

Accept-Language is a client supplied header detailing what locales the
client user prefers and the weight of the preference. It's
mostly useful to detect the default language of the site.

Modifications:

Added method `acceptLanguages()` and `selectLocale(supportedLocales)` to
`RequestHeaderGetters`. `acceptLanguages()` is lazy and cached.

Added `acceptLanguages(acceptedLanguages)` to `RequestHeadersBuilder`.

Closes #3179