Review comments by @trustin - solution for Authorizer chaining problem:
The AuthService invokes the first Authorizer form the chain. If it succeeds, it terminates the chain and invokes associated successHandler. If successHandler not defined for given Authorizer, it will invoke default/common successHandler. If the Authorizer fails (returns false), the AuthService tries the next Authorizer, and so on. At the end of the chain, if none of the Authroizers returned success, the AuthService selects the Authorizer to handle the failureHandler (LAST_WITH_HANDLER - the last one that has specific failureHandler defined). The AuthService always engages matching failureHandler from the selected Authorizer. If none of the Authorizers define a failureHandler, the AuthService selects engages default/common failureHandler.

Summary of the interface changes:

- AuthService has successHandler and failureHandler as before - these play a role of default/common handlers.
- AuthService has LAST_WITH_HANDLER Authorizer selection strategy hardcoded.
- Authorizer has failureHandler and successHandler set to NULL by default, as before.
- extracted Authorizer chaining logic into a dedicated class AuthorizerChain;
- extracted HttpRequest authorization logic into a separate class DelegatingHttpRequestAuthorizer.

This solution is fully backward compatible and the new changes mostly applied to AuthService#serve() and AuthServiceBuilder.