Cancel a `ArmeriaServerCall` when an `HttpRequest` is closed.
Motivation:

An `HttpRequest` is canceled by remote peer and an `HttpResponseWriter`
is closed, the `ArmeriaServerCall` should also be closed with
`Status.CANCELLED`. However, there is a race condition between
`response.write()` and `call.close()`.
The closing signal progragates to `ArmeriaServerCall` using `res.whenComplete()`.
The cancel action is not directly invoked by the signal. It is submitted to
an event loop and rescheduled.
https://github.com/line/armeria/blob/480321970111ac6249775199fedadcc71b862e84/grpc/src/main/java/com/linecorp/armeria/server/grpc/ArmeriaServerCall.java#L195-L204

Modifications:

- Directly try to cancel a server call when a response is closed.
- Cancel a server call when a `HttpResponseWriter` is closed already.

Result:

- You no longer see a `ClosedStreamException` when a gRPC call is cancelled
  by a client.
- Fixes #3452