Fix a bug where duplicated parameter specifications of a data class in DocService
Motivation:

The request parameter specification of a data class is shown with duplicated field information.
See #3454 for the detailed error case.
The `copy` method of data class is regarded as a redudent resolver.
`AnnotatedBeanFactoryRegistry` already has a detecting logic for duplication,
It only checks if the size of parameteters is equal to 1.
However if all parameters are redudent, there is no problem to skip the method.

Modifications:

- Do not add a annotated method resolver if all parameteres are redundent

Result:

- You no longer see reduedent parameters in DocService when Kotlin data class is used
- Fixes #3454