Correctly sends "Connection: close" header on disallowed HTTP/1 requests
Motivation:

When translating HTTP/2 headers to HTTP/1 headers,
Armeria filter out disallowed list from the headers.
For example, `"Connection"` is prohibited to use in HTTP/2
https://tools.ietf.org/html/rfc7540#section-8.1.2.2
A "Connection: close" header speficied for closing a connection is removed by the filter.
https://github.com/line/armeria/blob/32eabd2eb1f1be73f03a6425fa193631064cf2db/core/src/main/java/com/linecorp/armeria/internal/common/ArmeriaHttpUtil.java#L156
As a result, the client releases the inactive connection to the pool,
though the connection is closed by the server.

Modifications:

- Destory `KeepAliveHandler` on a failed request so that
  a "Connection: close" header correctly is included to the response headers.

Result:

Armeria server now correctly sends `"Connection: close"` header on an invalid request.