Add variants of DefaultContextAware*ExecutorService that propagate caller's RequestContext (#3483)
The current `DefaultContextAwareExecutorService` and `DefaultContextAwareScheduledExecutorService` run callbacks with a fixed `RequestContext`, meaning user code must call `RequestContext#makeContextAware` on each request and pass the returned `ExecutorService` through to the point of use. This can be cumbersome, particularly when interoperating with library code that is not generally ameria-aware and expects an `ExecutorService` to be wired up "statically" (for example a database-access service class that runs query result callbacks on a supplied executor service).

This PR offers parallel implementations that call `RequestContext.current()` internally at task-submit time, meaning that any tasks *submitted* by a request-context-aware thread will have that request context propagated to those tasks. These implementations may be more easy to use with library code, or in any scenario where an `ExecutorService` needs to be reused across multiple requests.